[FIX] create image util
